pyproject.toml
[project]
name = "sentio"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "cryptography>=46.0.4",
    "django>=6.0.1",
    "django-allauth>=65.14.0",
    "django-crispy-forms>=2.5",
    "djangorestframework>=3.16.1",
    "djangorestframework-stubs>=3.16.7",
    "lancedb>=0.27.1",
    "numpy>=2.4.1",
    "pandas>=3.0.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.2",
]
